parse http proxy port as int (more robust if port is given as string)

+++ files/groovy/setup_http_proxy.groovy
@@ -6,7 +6,7 @@ parsed_args = new JsonSlurper().parseText(args)
 core.removeHTTPProxy()
 if (parsed_args.with_http_proxy) {
     if (parsed_args.http_proxy_username) {
-        core.httpProxyWithBasicAuth(parsed_args.http_proxy_host, parsed_args.http_proxy_port, parsed_args.http_proxy_username, parsed_args.http_proxy_password)
+        core.httpProxyWithBasicAuth(parsed_args.http_proxy_host, parsed_args.http_proxy_port as int, parsed_args.http_proxy_username, parsed_args.http_proxy_password)
     } else {
         core.httpProxy(parsed_args.http_proxy_host, parsed_args.http_proxy_port as int)
     }
@@ -15,7 +15,7 @@ if (parsed_args.with_http_proxy) {
 core.removeHTTPSProxy()
 if (parsed_args.with_https_proxy) {
     if (parsed_args.https_proxy_username) {
-        core.httpsProxyWithBasicAuth(parsed_args.https_proxy_host, parsed_args.https_proxy_port, parsed_args.http_proxy_username, parsed_args.https_proxy_password)
+        core.httpsProxyWithBasicAuth(parsed_args.https_proxy_host, parsed_args.https_proxy_port as int, parsed_args.http_proxy_username, parsed_args.https_proxy_password)
     } else {
         core.httpsProxy(parsed_args.https_proxy_host, parsed_args.https_proxy_port as int)
     }
